Added new ML
- Added JVM target
- Add logic for saving and loading weights in the following modules:
  - LayerParameters (base class of all MLP layers)
  - MultiHeadAttentionParameters
- Add generic logic to accept any number of multi-hot features while computing an embedding
- Changed LernaEmbeddingsMap
  - From Any to Int in order to make the serialization easier
  - Removed some printouts, Verbose mode of trainer had erroneous outputs
- Changed Double to Float

+++ advancedml/src/commonMain/kotlin/com/kotlinnlp/simplednn/core/arrays/AugmentedArray.kt
@@ -185,11 +185,11 @@ open class AugmentedArray<NDArrayType : NDArray<NDArrayType>>(size: Int) : Activ
    * If [errors] are still not initialized a new [NDArray] is created.
    *
    * @param a the first factor, as [DenseNDArray]
-   * @param b the second factor, as Double number
+   * @param b the second factor, as Float number
    *
    * @return the errors of this [AugmentedArray]
    */
-  fun assignErrorsByProd(a: DenseNDArray, b: Double): DenseNDArray {
+  fun assignErrorsByProd(a: DenseNDArray, b: Float): DenseNDArray {
 
     require(a.length == this.size) { "Invalid arrays size" }
