+++ advancedml/src/commonMain/kotlin/com/kotlinnlp/simplednn/core/arrays/DistributionArray.kt
@@ -30,8 +30,8 @@ class DistributionArray(values: DenseNDArray) : Norm1Array<DenseNDArray>(values)
      */
     fun uniform(length: Int): DistributionArray {
 
-      val initValue = 1.0 / length
-      val values = DoubleArray(size = length, init = { initValue })
+      val initValue = 1.0f / length
+      val values = FloatArray(size = length, init = { initValue })
 
       return DistributionArray(values = DenseNDArrayFactory.arrayOf(values))
     }
@@ -56,7 +56,7 @@ class DistributionArray(values: DenseNDArray) : Norm1Array<DenseNDArray>(values)
 
   init {
     require((0 until this.values.length).all{ i -> this.values[i] in 0.0 .. 1.0}) { "Required 0 <= value[i] <= 1.0" }
-    require(equals(this.values.sum(), 1.0, tolerance = 1.0e-08)) { "Values sum must be equal to 1.0" }
+    require(equals(this.values.sum(), 1.0f, tolerance = 1.0e-08f)) { "Values sum must be equal to 1.0" }
     require(this.values.columns == 1) { "Values must be a column vector" }
   }
 
@@ -67,7 +67,7 @@ class DistributionArray(values: DenseNDArray) : Norm1Array<DenseNDArray>(values)
    */
   override fun assignValues(values: NDArray<*>) {
     require((0 until this.values.length).all{ i -> this.values[i] in 0.0 .. 1.0}) { "Required 0 <= value[i] <= 1.0" }
-    require(equals(values.sum(), 1.0, tolerance = 1.0e-08)) { "Values sum must be equal to 1.0" }
+    require(equals(values.sum(), 1.0f, tolerance = 1.0e-08f)) { "Values sum must be equal to 1.0" }
 
     this.values.assignValues(values)
   }
